eTactileKit - Release v1.1

+++ 2_Firmware/ETactileKit/lib/ADC_DAC/adc_dac.h
@@ -12,7 +12,9 @@ void initADC_DAC();
 
 /******************************************************************************/
 /* DA output by AD5452 and AD input by AD7276(SPI)                            */
+/* THe SPI MOSI is connected to the DAC and MISO is from the ADC output       */
+/* This eliminates the requirement for having seperate communication channels */
 /******************************************************************************/
-short DAAD(int DA);
+int DAAD(int DA);
 
 #endif //ADC_DAC_H
